Apply comments.
